requirements.txt
flask==3.0.0
dash==2.14.2
pandas==2.1.4
seaborn==0.13.0
matplotlib==3.8.2
python-dateutil==2.8.2
